// src/app/users/domain/models/ERole.ts
export enum ERole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  GLOBAL_MANAGER = 'GLOBAL_MANAGER',
  MANAGER_COMMERCIAL = 'MANAGER_COMMERCIAL',
  MANAGER_PRODUCTION = 'MANAGER_PRODUCTION',
  MANAGER_DISPATCH = 'MANAGER_DISPATCH',
  MANAGER_BILLING = 'MANAGER_BILLING',
  USER_COMMERCIAL = 'USER_COMMERCIAL',
  USER_PRODUCTION = 'USER_PRODUCTION',
  USER_DISPATCH = 'USER_DISPATCH',
  USER_BILLING = 'USER_BILLING',
  USER = 'USER',
}